# api.yaml
openapi: 3.0.0
info:
  title: Placement Portal API
  version: 1.0.0
  description: Campus placement management system API
servers:
  - url: http://localhost:5000/api
paths:
  /auth/login:
    post:
      summary: User authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
                  user_id:
                    type: integer
  /auth/register/student:
    post:
      summary: Student registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
                roll_number:
                  type: string
                branch:
                  type: string
                cgpa:
                  type: number
                year:
                  type: integer
      responses:
        '201':
          description: Student registered successfully
  /admin/dashboard:
    get:
      summary: Admin dashboard statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_students:
                    type: integer
                  total_companies:
                    type: integer
                  total_drives:
                    type: integer
  /drives:
    get:
      summary: List approved placement drives
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of drives
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    job_title:
                      type: string
                    company_name:
                      type: string
                    min_cgpa:
                      type: number
                    application_deadline:
                      type: string
                      format: date-time
  /drives/{id}/apply:
    post:
      summary: Apply to placement drive
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Application submitted successfully
        '400':
          description: Eligibility requirements not met
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT